requirements.txt
requests
pandas
schedule
python-dotenv
smartapi-python
python-telegram-bot